route fixed, logged in persists at page refresh
