+++ client/src/pages/NoPage.jsx
@@ -1,0 +1,12 @@
+import Header from "../components/Header";
+
+const NoPage = () => {
+  return (
+    <div>
+      <Header />
+      <h1>Error: 404 Page Not Found</h1>
+    </div>
+  )
+}
+
+export default NoPage;
